requirements.txt
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
markupsafe==2.1.3
